+++ backend/src/baserow/config/settings/base.py
@@ -203,6 +203,7 @@ SPECTACULAR_SETTINGS = {
         {"name": "Database table grid view"},
         {"name": "Database table rows"},
         {"name": "Database tokens"},
+        {"name": "Admin"},
     ],
 }