+++ premium/backend/src/baserow_premium/api/admin/users/errors.py
@@ -1,0 +1,20 @@
+from rest_framework.status import HTTP_400_BAD_REQUEST
+
+
+USER_ADMIN_CANNOT_DEACTIVATE_SELF = (
+    "USER_ADMIN_CANNOT_DEACTIVATE_SELF",
+    HTTP_400_BAD_REQUEST,
+    "You cannot de-activate or un-staff yourself.",
+)
+
+USER_ADMIN_CANNOT_DELETE_SELF = (
+    "USER_ADMIN_CANNOT_DELETE_SELF",
+    HTTP_400_BAD_REQUEST,
+    "You cannot delete yourself.",
+)
+
+USER_ADMIN_UNKNOWN_USER = (
+    "USER_ADMIN_UNKNOWN_USER",
+    HTTP_400_BAD_REQUEST,
+    "Unknown user supplied.",
+)