+++ premium/web-frontend/modules/baserow_premium/components/admin/groups/contexts/HiddenUsersContext.vue
@@ -1,0 +1,13 @@
+<script>
+import HiddenGroupsContext from '@baserow_premium/components/admin/users/contexts/HiddenGroupsContext'
+
+export default {
+  name: 'HiddenUsersContext',
+  extends: HiddenGroupsContext,
+  data() {
+    return {
+      nameKey: 'email',
+    }
+  },
+}
+</script>