+++ premium/web-frontend/modules/baserow_premium/pages/admin/users.vue
@@ -1,0 +1,13 @@
+<template>
+  <UsersAdminTable></UsersAdminTable>
+</template>
+
+<script>
+import UsersAdminTable from '@baserow_premium/components/admin/users/UsersAdminTable'
+
+export default {
+  components: { UsersAdminTable },
+  layout: 'app',
+  middleware: 'staff',
+}
+</script>