+++ web-frontend/modules/core/adminTypes.js
@@ -28,7 +28,7 @@ export class AdminType extends Registerable {
    * The order value used to sort admin types in the sidebar menu.
    */
   getOrder() {
-    return 1
+    return 0
   }
 
   getRouteName() {
@@ -81,5 +81,9 @@ export class SettingsAdminType extends AdminType {
 
   getRouteName() {
     return 'admin-settings'
+  }
+
+  getOrder() {
+    return 9999
   }
 }